+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 
 ## 2.0 (work in progress)
 
+- Support `cc` and `bcc` parameters
 - Replace deprecated `openURL(_)` with `open(_, options: , completionHandler:)`
+- Update `ThirdPartyMailClient.composeURL(...)` with default parameters values
+- Rename `ThirdPartyMailClient` `clients()` to `clients`
 - Update `ThirdPartyMailer` functions with completion handler instead of boolean return value
+- Rename `ThirdPartyMailer` `application(_, isMailClientAvailable:)` to `isMailClientAvailable(_, with:)`, with default parameters values
+- Rename `ThirdPartyMailer` `application(_, openMailClient:, completionHandler:)` to `open(_, with:, completionHandler:)`, with default parameters values
+- Rename `ThirdPartyMailer` `application(_, openMailClient:, recipient:, subject:, body:, completionHandler:)` to `openCompose(_, recipient:, subject:, body:, cc:, bcc:, with:, completionHandler:)`, with default parameters values
 - Require iOS 10